Backend/requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
python-socketio[client]>=5.10.0
